Move exception to appropriate package
