+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/cloud/azure/batchpool/DeleteAzureBatchPoolStep.java
@@ -6,7 +6,7 @@ import bio.terra.stairway.StepResult;
 import bio.terra.stairway.StepStatus;
 import bio.terra.stairway.exception.RetryException;
 import bio.terra.workspace.app.configuration.external.AzureConfiguration;
-import bio.terra.workspace.common.utils.ManagementExceptionUtils;
+import bio.terra.workspace.common.exception.AzureManagementExceptionUtils;
 import bio.terra.workspace.generated.model.ApiAzureLandingZoneDeployedResource;
 import bio.terra.workspace.service.crl.CrlService;
 import bio.terra.workspace.service.iam.AuthenticatedUserRequest;
@@ -76,8 +76,8 @@ public class DeleteAzureBatchPoolStep implements Step {
           batchAccountName.get());
       return StepResult.getStepResultSuccess();
     } catch (ManagementException e) {
-      if (ManagementExceptionUtils.isExceptionCode(
-          e, ManagementExceptionUtils.RESOURCE_NOT_FOUND)) {
+      if (AzureManagementExceptionUtils.isExceptionCode(
+          e, AzureManagementExceptionUtils.RESOURCE_NOT_FOUND)) {
         logger.info(
             "Azure Batch Pool '{}' in batch account '{}' already deleted",
             resource.getId(),
