+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/cloud/azure/storageContainer/CreateAzureStorageContainerStep.java
@@ -7,7 +7,7 @@ import bio.terra.stairway.StepResult;
 import bio.terra.stairway.StepStatus;
 import bio.terra.stairway.exception.RetryException;
 import bio.terra.workspace.app.configuration.external.AzureConfiguration;
-import bio.terra.workspace.common.utils.ManagementExceptionUtils;
+import bio.terra.workspace.common.exception.AzureManagementExceptionUtils;
 import bio.terra.workspace.service.crl.CrlService;
 import bio.terra.workspace.service.resource.controlled.cloud.azure.storageContainer.resourcemanager.data.CreateStorageContainerRequestData;
 import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.ControlledResourceKeys;
@@ -112,8 +112,8 @@ public class CreateAzureStorageContainerStep implements Step {
           .storageAccounts()
           .getByResourceGroup(azureCloudContext.getAzureResourceGroupId(), storageAccountName);
     } catch (ManagementException ex) {
-      if (ManagementExceptionUtils.isExceptionCode(
-          ex, ManagementExceptionUtils.RESOURCE_NOT_FOUND)) {
+      if (AzureManagementExceptionUtils.isExceptionCode(
+          ex, AzureManagementExceptionUtils.RESOURCE_NOT_FOUND)) {
         logger.warn(
             "Deletion of the storage container is not required. Parent storage account does not exist. {}",
             storageAccountName);
@@ -134,8 +134,8 @@ public class CreateAzureStorageContainerStep implements Step {
               storageAccountName,
               resource.getStorageContainerName());
     } catch (ManagementException ex) {
-      if (ManagementExceptionUtils.isExceptionCode(
-          ex, ManagementExceptionUtils.CONTAINER_NOT_FOUND)) {
+      if (AzureManagementExceptionUtils.isExceptionCode(
+          ex, AzureManagementExceptionUtils.CONTAINER_NOT_FOUND)) {
         logger.warn(
             "Deletion of the storage container is not required. Storage container does not exist. {}",
             resource.getStorageContainerName());
