+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/cloud/azure/vm/CreateAzureVmStep.java
@@ -7,10 +7,10 @@ import bio.terra.cloudres.azure.resourcemanager.compute.data.CreateVirtualMachin
 import bio.terra.stairway.*;
 import bio.terra.stairway.exception.RetryException;
 import bio.terra.workspace.app.configuration.external.AzureConfiguration;
-import bio.terra.workspace.common.utils.AzureManagementException;
+import bio.terra.workspace.common.exception.AzureManagementException;
+import bio.terra.workspace.common.exception.AzureManagementExceptionUtils;
 import bio.terra.workspace.common.utils.AzureVmUtils;
 import bio.terra.workspace.common.utils.FlightUtils;
-import bio.terra.workspace.common.utils.ManagementExceptionUtils;
 import bio.terra.workspace.db.ResourceDao;
 import bio.terra.workspace.generated.model.ApiAzureVmCreationParameters;
 import bio.terra.workspace.service.crl.CrlService;
@@ -149,7 +149,7 @@ public class CreateAzureVmStep implements Step {
       // Stairway steps may run multiple times, so we may already have created this resource. In all
       // other cases, surface the exception and attempt to retry.
       return switch (e.getValue().getCode()) {
-        case ManagementExceptionUtils.CONFLICT -> {
+        case AzureManagementExceptionUtils.CONFLICT -> {
           logger.info(
               "Azure Vm {} in managed resource group {} already exists",
               resource.getVmName(),
@@ -157,7 +157,7 @@ public class CreateAzureVmStep implements Step {
           yield StepResult.getStepResultSuccess();
         }
 
-        case ManagementExceptionUtils.RESOURCE_NOT_FOUND -> {
+        case AzureManagementExceptionUtils.RESOURCE_NOT_FOUND -> {
           logger.error(
               "Either the disk, ip, or network passed into this createVm does not exist "
                   + String.format(
@@ -170,14 +170,14 @@ public class CreateAzureVmStep implements Step {
               StepStatus.STEP_RESULT_FAILURE_FATAL, new AzureManagementException(e));
         }
 
-        case ManagementExceptionUtils.VM_EXTENSION_PROVISIONING_ERROR -> {
+        case AzureManagementExceptionUtils.VM_EXTENSION_PROVISIONING_ERROR -> {
           logger.error("Error provisioning VM extension");
           yield new StepResult(
               StepStatus.STEP_RESULT_FAILURE_FATAL, new AzureManagementException(e));
         }
 
         default -> new StepResult(
-            ManagementExceptionUtils.maybeRetryStatus(e), new AzureManagementException(e));
+            AzureManagementExceptionUtils.maybeRetryStatus(e), new AzureManagementException(e));
       };
     }
     return StepResult.getStepResultSuccess();
